+++ MiClaseRemota2/src/MiInterfazRemota.java
@@ -3,6 +3,6 @@ import java.rmi.Remote;
 import java.rmi.RemoteException;
 
 public interface MiInterfazRemota extends Remote {
-    long miMetodo1(Integer[] array) throws RemoteException;
+    long miMetodo1(Long[] array) throws RemoteException;
     void methodReturn() throws RemoteException;
 }
